Allow to disable offlining of non-boot CPUs
On my Cherrytrail tablet I've been trying to achieve the S0Ix deep sleep
states. After a bit of twiddling I found out that I can achieve the deep
sleep states but I need to disable the power off of the secondary cores
otherwise the resume from sleep causes the device to crash.

The patch uses a property to determine whether to power off the non-boot
CPUs on sleeping.